Double message event
The deleted code cause a double delete event, meaning that the same message got 2 delete events. In the event you would want to the bot to delete a message after for example posting a command it will cause a double call on the delete event causing the message signalled as delete 2 times.

Compiled and tested on a bot that logs deleted messages. After removing said code the double logging of the deleted message no longer occurred.