Store websocket and heartbeatTimer in AtomicReferences to prevent memory leaks and use the central scheduler for the heartbeating instead of an own Timer
Different threads use those variables, not concurrently, but still the different ReadingThread instances of the different web sockets on reconnect can and will have different values for those variables if access is not synchronized. This means that there will remain hard references and you leak memory.
By using AtomicReferences for those variables, the leak is fixed and the instances can get garbage collected properly.

Furthermore now the central scheduler is used for the hearbeating instead of creating an own Timer.