Fix the delay from a failed resume until the subsequent identify
The identify-rate-limit semaphore was held for the connect call that tries to resume and was not released on resume, this caused Javacord to wait until the starvation protector task released the semaphore and then the identify call was done.
Now when a resume is tried, the semaphore gets released immediately as it should only preven concurrent identify calls and by that, the reconnect delay is working as intended and not always about 10 seconds longer than expected.