Introduce Permissionable interface.
Permissionable is a marker interface for discord entities that can be
granted permissions. While internally it is most efficient to store only
the entity ids in separate maps in order to maintain the type information,
user-facing APIs will accept a generic permissionable object as to not
leak implementation details.

In situations where user permissions are calculated, the method will
still only accept a User object.