Catch Throwable instead of Exception when handling packets
Without this change things like NoSuchMethodError due to incompatible Jackson version
are just swallowed and the bot might just not work properly.
This for example happened when a server text channel inside a category was constructed
from the GUILD_CREATE package and Jackson version where the used method was not yet
available. The bot just didn't work on those servers but worked fine on ones that only
have text channels outside categories or in private messages.