Improve performance of getters for channels  …
Added a lookup table to map the channel id to the object. This
improved the performance of the getXyzChannelById(...) methods
significantly. As these methods are used internally, this is a
very important change for Javacord's overall performance.

The getXyzChannels() methods were also improved slightly, but there's
still a lot of room for improvements. As these are not used internally,
they are also not as important as the getters by id.