Fix AudioConnection#setSpeaking(boolean) method
When calling the AudioConnectionImpl#setSpeakingFlags(...) method, the
method checks if the flags did change and if yes, it sends an update
to the voice websocket.
The #set[Priority]Speaking(boolean) method however did modify the
speakingFlags enum set directly before calling #setSpeakingFlags(...).
Because of this the check always yielded false (no change) and thus
it never send something to the voice websocket.