Fix #waitForUsersOnStartup()
If you have the startup wait for users enabled but not the
GUILD_PRESENCES intent, your bot will be unable to load all members
of non-large servers and mark them as unavailable. It also causes
unnecessary guild member requests if you have the user wait disabled
but the cache enabled.