fix: switch "nick" to "global_name" where applicable
The Discord API phased out `nick` as the field for the nickname for `global_name`. The only thing that I kept `nick` is the updater as it is still indicated to be `nick` in the PATCH request. All data structures, methods, and so on keep the phrase Nickname as it would likely be an unnecessary breaking change.